refactored file structure to enable jest testing
