+++ package.json
@@ -4,7 +4,9 @@
   "main": "index.js",
   "scripts": {
     "server": "nodemon src/index.js",
-    "heroku-postbuild": "npm install && npx prisma migrate deploy && npx prisma generate"
+    "heroku-postbuild": "npm install && npx prisma migrate deploy && npx prisma generate",
+    "test": "jest --runInBand",
+    "test:watch": "jest --watch"
   },
   "repository": {
     "type": "git",
@@ -29,6 +31,8 @@
     "nodemon": "^3.1.10"
   },
   "devDependencies": {
-    "prisma": "^6.7.0"
+    "jest": "^29.7.0",
+    "prisma": "^6.7.0",
+    "supertest": "^7.1.0"
   }
 }
